feat(rules): `new-for-builtins` with crazy restriction for `new Object(123)`
the `new Object(123)` generates a boxed version of a Number, the most confusing
case is `new Boolean(false)` and `new Object(false)` - both give a boxed version
of a `false` (which evaluates to `true` in conditions because of its
box-wrapper), but since today there are no need to use a "boxed bool"
or any other primitive - it is prohibited by a crazy setup with
a custom `no-restricted-syntax`. Probably, one day somebody replaces this
rule with a custom new rule.